requirements.txt
Kivy
requirements
dnspython